backend/requirements.txt
# --- Core Frameworks ---
langchain>=0.3.0
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=1.0.0

# --- Google HAI-DEF & MedGemma Integration ---
# Using the consolidated SDK for Vertex AI access
langchain-google-vertexai>=2.0.0
langchain-google-genai>=2.0.0
google-cloud-aiplatform>=1.70.0

# --- Vector Database & Embeddings ---
chromadb>=0.5.0
langchain-huggingface>=1.2.0
sentence-transformers>=3.0.0
sentencepiece>=0.2.0         # Required by SigLIP/MedSigLIP tokenizer
torch>=2.0.0
transformers>=4.40.0         # For MedSigLIP/SigLIP text encoder (EMBEDDING_MODEL=medsiglip)

# --- Web Interface & API ---
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
requests>=2.32.0
python-dotenv>=1.0.0
PyPDF2>=3.0.1

# --- Data Processing ---
lxml>=5.0.0      # Required for eCFR XML parsing
pandas>=2.2.0

# --- Evaluation & Visualisation ---
matplotlib>=3.8.0   # Retrieval evaluation plots (test/evaluate_retrieval.py)